# docs/swagger.yaml
basePath: /api
definitions:
  http.APIError:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      message:
        example: Invalid input data
        type: string
    type: object
  http.APIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/http.APIError'
      success:
        example: false
        type: boolean
    type: object
  http.APIResponse:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  http.ActivateAccountRequest:
    properties:
      token:
        example: activation-token-here
        type: string
    required:
    - token
    type: object
  http.ChangePasswordRequest:
    properties:
      new_password:
        example: newpassword123
        minLength: 6
        type: string
      token:
        example: reset-token-here
        type: string
    required:
    - new_password
    - token
    type: object
  http.ForgotPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
  http.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  http.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: support@kairo-anchor.com
    name: API Support
  description: Authentication API with JWT tokens for Kairo Anchor application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Kairo Anchor Authentication API
  version: "1.0"
paths:
  /auth/activate:
    post:
      consumes:
      - application/json
      description: Activate user account using activation token from email
      parameters:
      - description: Activate Account Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ActivateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.APIErrorResponse'
      summary: Activate user account
      tags:
      - auth
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change password using reset token from email
      parameters:
      - description: Change Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.APIErrorResponse'
      summary: Change password
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email to user
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.APIErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.APIErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/profile:
    get:
      consumes:
      - application/json
      description: Get authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account and send activation email
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.APIErrorResponse'
      summary: Register a new user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"